+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix LocalVideo not rendering on initial attempt
 - Fix inconsistent snapshot for EditableChatBubble
+- [Docs] Fix mdx style rendering
 
 ### Added