+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Allow property passthough to IconButton in NavbarItem
+- Render roster without waiting for getAtendee callback
 
 ### Removed