+++ CHANGELOG.md
@@ -11,6 +11,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+### Added
+
+### Changed
+
+- Render roster without waiting for getAtendee callback
+
+### Removed
+
+## [2.0.1] - 2020-2-11
+
+### Fixed
+
 - Fix LocalVideo not rendering on initial attempt
 - Fix inconsistent snapshot for EditableChatBubble
 - [Docs] Fix mdx style rendering
@@ -23,7 +35,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Allow property passthough to IconButton in NavbarItem
-- Render roster without waiting for getAtendee callback
 
 ### Removed